Ability Effect Modifications
Modified the Ability Effect construction, so we can start creating
unique effects like DoT, Slow, etc.
This required some changes to be done to how we utilized the
CombatManager. I refactored the code so that the CombatManager now
manages the different ability Effects, which then casts the effect
through a public method within the effect. I refactored the code so that
we do not use SendMessage method anymore to improve efficiency. I also
created a generic health class. This commit is mostly to allow me to
change to my laptop so I can continue working. Still need to
cleanup/comment that code. The Effect is still not applied to the target
as well.